Physically organize function decls.
