+++ src/pybind/wrapper.cpp
@@ -5,6 +5,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#include "wrapper.hpp"
+
 #include "codegen/codegen_naming.hpp"
 #include "pybind/pyembed.hpp"
